+++ taiga/projects/tasks/permissions.py
@@ -15,7 +15,8 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from taiga.base.api.permissions import (TaigaResourcePermission, HasProjectPerm,
-                                        IsProjectOwner, AllowAny, IsSuperUser)
+                                        IsAuthenticated, IsProjectOwner, AllowAny,
+                                        IsSuperUser)
 
 
 class TaskPermission(TaigaResourcePermission):
@@ -30,3 +31,12 @@ class TaskPermission(TaigaResourcePermission):
     csv_perms = AllowAny()
     bulk_create_perms = HasProjectPerm('add_task')
     bulk_update_order_perms = HasProjectPerm('modify_task')
+    upvote_perms = IsAuthenticated() & HasProjectPerm('view_tasks')
+    downvote_perms = IsAuthenticated() & HasProjectPerm('view_tasks')
+
+
+class TaskVotersPermission(TaigaResourcePermission):
+    enought_perms = IsProjectOwner() | IsSuperUser()
+    global_perms = None
+    retrieve_perms = HasProjectPerm('view_tasks')
+    list_perms = HasProjectPerm('view_tasks')