+++ taiga/projects/tasks/serializers.py
@@ -27,12 +27,14 @@ from taiga.projects.milestones.validators import SprintExistsValidator
 from taiga.projects.tasks.validators import TaskExistsValidator
 from taiga.projects.notifications.validators import WatchersValidator
 from taiga.projects.serializers import BasicTaskStatusSerializerSerializer
+from taiga.projects.votes.mixins.serializers import VotedResourceSerializerMixin
+
 from taiga.users.serializers import UserBasicInfoSerializer
 
 from . import models
 
 
-class TaskSerializer(WatchersValidator, serializers.ModelSerializer):
+class TaskSerializer(WatchersValidator, VotedResourceSerializerMixin, serializers.ModelSerializer):
     tags = TagsField(required=False, default=[])
     external_reference = PgArrayField(required=False)
     comment = serializers.SerializerMethodField("get_comment")