+++ taiga/projects/userstories/permissions.py
@@ -30,3 +30,12 @@ class UserStoryPermission(TaigaResourcePermission):
     csv_perms = AllowAny()
     bulk_create_perms = IsAuthenticated() & (HasProjectPerm('add_us_to_project') | HasProjectPerm('add_us'))
     bulk_update_order_perms = HasProjectPerm('modify_us')
+    upvote_perms = IsAuthenticated() & HasProjectPerm('view_us')
+    downvote_perms = IsAuthenticated() & HasProjectPerm('view_us')
+
+
+class UserStoryVotersPermission(TaigaResourcePermission):
+    enought_perms = IsProjectOwner() | IsSuperUser()
+    global_perms = None
+    retrieve_perms = HasProjectPerm('view_us')
+    list_perms = HasProjectPerm('view_us')