+++ tests/integration/test_vote_userstories.py
@@ -23,11 +23,11 @@ from .. import factories as f
 pytestmark = pytest.mark.django_db
 
 
-def test_upvote_issue(client):
+def test_upvote_user_story(client):
     user = f.UserFactory.create()
-    issue = f.create_issue(owner=user)
-    f.MembershipFactory.create(project=issue.project, user=user, is_owner=True)
-    url = reverse("issues-upvote", args=(issue.id,))
+    user_story = f.create_userstory(owner=user)
+    f.MembershipFactory.create(project=user_story.project, user=user, is_owner=True)
+    url = reverse("userstories-upvote", args=(user_story.id,))
 
     client.login(user)
     response = client.post(url)
@@ -35,11 +35,11 @@ def test_upvote_issue(client):
     assert response.status_code == 200
 
 
-def test_downvote_issue(client):
+def test_downvote_user_story(client):
     user = f.UserFactory.create()
-    issue = f.create_issue(owner=user)
-    f.MembershipFactory.create(project=issue.project, user=user, is_owner=True)
-    url = reverse("issues-downvote", args=(issue.id,))
+    user_story = f.create_userstory(owner=user)
+    f.MembershipFactory.create(project=user_story.project, user=user, is_owner=True)
+    url = reverse("userstories-downvote", args=(user_story.id,))
 
     client.login(user)
     response = client.post(url)
@@ -47,12 +47,12 @@ def test_downvote_issue(client):
     assert response.status_code == 200
 
 
-def test_list_issue_voters(client):
+def test_list_user_story_voters(client):
     user = f.UserFactory.create()
-    issue = f.create_issue(owner=user)
-    f.MembershipFactory.create(project=issue.project, user=user, is_owner=True)
-    f.VoteFactory.create(content_object=issue, user=user)
-    url = reverse("issue-voters-list", args=(issue.id,))
+    user_story = f.create_userstory(owner=user)
+    f.MembershipFactory.create(project=user_story.project, user=user, is_owner=True)
+    f.VoteFactory.create(content_object=user_story, user=user)
+    url = reverse("userstory-voters-list", args=(user_story.id,))
 
     client.login(user)
     response = client.get(url)
@@ -60,12 +60,12 @@ def test_list_issue_voters(client):
     assert response.status_code == 200
     assert response.data[0]['id'] == user.id
 
-def test_get_issue_voter(client):
+def test_get_userstory_voter(client):
     user = f.UserFactory.create()
-    issue = f.create_issue(owner=user)
-    f.MembershipFactory.create(project=issue.project, user=user, is_owner=True)
-    vote = f.VoteFactory.create(content_object=issue, user=user)
-    url = reverse("issue-voters-detail", args=(issue.id, vote.user.id))
+    user_story = f.create_userstory(owner=user)
+    f.MembershipFactory.create(project=user_story.project, user=user, is_owner=True)
+    vote = f.VoteFactory.create(content_object=user_story, user=user)
+    url = reverse("userstory-voters-detail", args=(user_story.id, vote.user.id))
 
     client.login(user)
     response = client.get(url)
@@ -73,13 +73,14 @@ def test_get_issue_voter(client):
     assert response.status_code == 200
     assert response.data['id'] == vote.user.id
 
-def test_get_issue_votes(client):
-    user = f.UserFactory.create()
-    issue = f.create_issue(owner=user)
-    f.MembershipFactory.create(project=issue.project, user=user, is_owner=True)
-    url = reverse("issues-detail", args=(issue.id,))
 
-    f.VotesFactory.create(content_object=issue, count=5)
+def test_get_user_story_votes(client):
+    user = f.UserFactory.create()
+    user_story = f.create_userstory(owner=user)
+    f.MembershipFactory.create(project=user_story.project, user=user, is_owner=True)
+    url = reverse("userstories-detail", args=(user_story.id,))
+
+    f.VotesFactory.create(content_object=user_story, count=5)
 
     client.login(user)
     response = client.get(url)
@@ -88,14 +89,14 @@ def test_get_issue_votes(client):
     assert response.data['votes'] == 5
 
 
-def test_get_issue_is_voted(client):
+def test_get_user_story_is_voted(client):
     user = f.UserFactory.create()
-    issue = f.create_issue(owner=user)
-    f.MembershipFactory.create(project=issue.project, user=user, is_owner=True)
-    f.VotesFactory.create(content_object=issue)
-    url_detail = reverse("issues-detail", args=(issue.id,))
-    url_upvote = reverse("issues-upvote", args=(issue.id,))
-    url_downvote = reverse("issues-downvote", args=(issue.id,))
+    user_story = f.create_userstory(owner=user)
+    f.MembershipFactory.create(project=user_story.project, user=user, is_owner=True)
+    f.VotesFactory.create(content_object=user_story)
+    url_detail = reverse("userstories-detail", args=(user_story.id,))
+    url_upvote = reverse("userstories-upvote", args=(user_story.id,))
+    url_downvote = reverse("userstories-downvote", args=(user_story.id,))
 
     client.login(user)